requirements.txt
cryptography==39.0.2
coverage==7.2.1
Django==4.1.7
django-cors-headers==3.4.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.4
mixer==7.2.2
phonenumbers==8.12.14
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
pytest-sugar==0.9.6
pytest-xdist==3.1.0
python-dotenv==1.0.0